Toss/Toss.Server/Models/Account/ChangePasswordCommandHandler.cs
using MediatR;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Identity;
using Microsoft.Extensions.Logging;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading;
using System.Threading.Tasks;
using Toss.Server.Controllers;
using Toss.Server.Extensions;
using Toss.Shared;
using Toss.Shared.Account;

namespace Toss.Server.Models.Account
{
    public class ChangePasswordCommandHandler : IRequestHandler<ChangePasswordCommand,CommandResult> 
    {
        private readonly UserManager<ApplicationUser> _userManager;
        private readonly SignInManager<ApplicationUser> _signInManager;
        private readonly ILogger _logger;
        private readonly IHttpContextAccessor _httpContextAccessor;

        public ChangePasswordCommandHandler(UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager, ILogger<AccountController> logger, IHttpContextAccessor httpContextAccessor)
        {
            _userManager = userManager;
            _signInManager = signInManager;
            _logger = logger;
            _httpContextAccessor = httpContextAccessor;
        }

        public async Task<CommandResult> Handle(ChangePasswordCommand request, CancellationToken cancellationToken)
        {
            var httpUser = _httpContextAccessor.HttpContext.User;
            var user = await _userManager.GetUserAsync(httpUser);
            if (user == null)
            {
                throw new ApplicationException($"Unable to load user with ID '{_userManager.GetUserId(httpUser)}'.");
            }
            if (string.IsNullOrEmpty(user.PasswordHash))
            {
                return new CommandResult("User", "User has no password set");
            }
            var changePasswordResult = await _userManager.ChangePasswordAsync(user, request.OldPassword, request.NewPassword);
            if (!changePasswordResult.Succeeded)
            {
                return new CommandResult(changePasswordResult.ToValidationErrorDictonary());
            }

            await _signInManager.SignInAsync(user, isPersistent: false);
            _logger.LogInformation("User changed their password successfully.");
            return CommandResult.Success();
        }
    }
}
